Create clone_check.py
This file contains a tool designed to check for the presence of a target sequence in sanger sequenced plasmids. Primarily this tool will be used to ensure that the target sequence has been properly inserted into the plasmid using a simple exact matching algorithm. This tool is not intended for alignment purposes.